requirements.txt
openai==2.5.0
flask==3.1.2
pandas==2.1.4
python-dotenv==1.1.1
requests==2.32.5
markdown==3.9
werkzeug==3.1.3